Create base sensor class and derivatives for C# and VB.NET
Create a new abstract base class to parse the report file generated by InspectCode and make the detected issues available within SonarQube if the corresponding rule of the repository is activated for the project's profile. Most of the base implementation has been extracted from the previous implementation of the CSharpSensor class. The derivatives for the languages C# and VisualBasic.NET provide language specific configuration data which is then used by the base implementation.
